consumer/requirements.txt

kafka-python==2.2.15
